Development ongoig - Session Form Refined again and other grids and forms Implemented Mobx Everywhere

+++ client/scripts/components/Account/AccountGrid.jsx
@@ -9,40 +9,61 @@ class AccountGrid extends Component {
     }
 
 	static defaultProps = {
-        accountsList : [],
+        entriesList : [],
         onDataUpdate: function() {},
         editItem: function(account_id){}
        
     }
 
-    render() {
-    	let source = {
+    componentWillMount() {
+        this.initDataAdapter()
+        // console.log(this.refs.jqxgrid)
+    }
+
+    componentDidUpdate() {
+        this.source.localdata = this.props.entriesList.slice()
+        this.dataAdapter.dataBind()
+        // this.refs.jqxgrid.refresh()
+        // this.refs.jqxgrid.updatebounddata('cells')
+    }
+
+    componentDidMount() {
+        // console.log("DID MOUnt")
+        window.grid = this.refs.jqxgrid
+        window.grid_data = this.props.entriesList.slice()
+        window.dataAdapter = this.dataAdapter
+    }
+
+    initDataAdapter() {
+         var datafields = [
+            { name: '_id', type: 'string' },
+            { name: 'account_name', type: 'string' },
+            { name: 'limit', type: 'number' },
+            { name: 'status', type: 'boolean' },
+            { name: 'hide', type: 'boolean' },
+        ];
+
+        this.source = {
             datatype: 'json',
-            datafields: [
-            	{ name: '_id', type: 'number' },
-                { name: 'account_name', type: 'string' },
-                { name: 'limit', type: 'number' },
-                { name: 'status', type: 'boolean' },
-                { name: 'hide', type: 'boolean' },
-            ],
+            datafields: datafields,
 
             id: '_id',
             // url: '/accounts',
-            localdata: this.props.accountsList,
+            localdata: this.props.entriesList.slice(),
 
-            updaterow: (rowid, rowdata, commit) => {
-                AccountHelper.update(rowdata.uid, {
-                    account_name: rowdata.account_name
-                }).then((res) => {
-                	this.props.onDataUpdate()
-                })
-                commit(true);
-            },
+            // updaterow: (rowid, rowdata, commit) => {
+            //     AccountHelper.update(rowdata.uid, {
+            //         account_name: rowdata.account_name
+            //     }).then((res) => {
+            //         this.props.onDataUpdate()
+            //     })
+            //     commit(true);
+            // },
         };
 
-        let dataAdapter = new $.jqx.dataAdapter(source);
+        this.dataAdapter = new $.jqx.dataAdapter(this.source);
 
-        let columns = [
+        this.columns = [
             {
                 text: 'Delete',
                 datafield: 'Delete',
@@ -50,14 +71,19 @@ class AccountGrid extends Component {
                 width: 50,
                 filterable: false,
                 cellsrenderer: () => {
-                    return 'Delete';
+                    return 'Del';
                 },
                 buttonclick: (row) => {
                     let dataRecord = this.refs.jqxgrid.getrowdata(row);
-                    console.log(dataRecord.uid)
-                    AccountHelper.delete(dataRecord.uid).then( (res) => {
-                        this.props.onDataUpdate()
-                    })
+                    // console.log(dataRecord.uid)
+                    var r = confirm("Are you sure!");
+                    if (r == true) {
+                        AccountHelper.delete(dataRecord.uid).then( (res) => {
+                            this.props.onDataUpdate()
+                        }).catch((err) => {
+                            toastr.error(err.response.data.message)
+                        })
+                    }
 
                 }
             },
@@ -72,23 +98,27 @@ class AccountGrid extends Component {
                 },
                 buttonclick: (row) => {
                     let dataRecord = this.refs.jqxgrid.getrowdata(row);
-                   this.props.editItem(dataRecord.uid)
+                    this.props.editItem(dataRecord.uid)
 
                 }
             },
-        	{ text: 'Id', datafield: '_id', width: 50 },
+            { text: 'Id', datafield: '_id', width: 50 },
             { text: 'Name', datafield: 'account_name', width: 230 },
             { text: 'Limit', datafield: 'limit', width: 100 },
             { text: 'Status', datafield: 'status', width: 50, columntype: 'checkbox' },
             { text: 'Hide', datafield: 'hide', width: 50, columntype: 'checkbox' },
         ];
+    }
 
+    render() {
         return (
             <div>
-         		 <JqxGrid ref="jqxgrid" width={"100%"} 
-                            height={600} source={dataAdapter} pageable={true} 
-                            sortable={true} altrows={true} enabletooltips={true} 
-                            editable={true} columns={columns} filterable={true} showfilterrow={true} />
+         		<JqxGrid ref="jqxgrid" key1={Math.random()}
+                   source={this.dataAdapter} columns={this.columns}
+                    width={"100%"} height={400} pageable={true} pagermode={'simple'} pagesize={1000}
+                    sortable={true} altrows={false} enabletooltips={true}
+                    editable={false} 
+                    filterable={false} showfilterrow={false} />
             </div>
         );
     }
